♻️ Refactor contract/item to use sol-kit
