+++ js/package.json
@@ -39,7 +39,7 @@
     "@project-serum/anchor": "^0.24.2",
     "@solana/spl-token": "^0.1.8",
     "@raindrop-studios/sol-command": "^0.4.0",
-    "@raindrop-studios/sol-kit": "^0.1.71",
+    "@raindrop-studios/sol-kit": "^0.1.72",
     "borsh": "^0.4.0",
     "commander": "^8.1.0",
     "form-data": "^4.0.0",
